requirements.txt
scikit-learn
scipy
python-bidi
arabic-reshaper
matplotlib
numpy
tensorflow
